+++ build.zig
@@ -37,6 +37,7 @@ pub fn build(b: *std.Build) void {
     const opts = b.addOptions();
     opts.addOption(bool, "time_profile", b.option(bool, "profile", "profile the time loading takes") orelse false);
     opts.addOption(bool, "dump_vpk", b.option(bool, "dumpvpk", "dump all vpk entries to text file") orelse false);
+    opts.addOption([]const u8, "commit_hash", b.option([]const u8, "commit_hash", "Embed git hash in build") orelse "unspecified");
     hammer_exe.root_module.addOptions("config", opts);
 
     // This declares intent for the executable to be installed into the