+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # Changelog for v0.0.4
+
+* Add help for shell commands
+* add --build flag to print build information